requirements.txt
dogpile.cache==0.6.4
et-xmlfile==1.0.1
flapjack-stack==1.0.0
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.9
orderedset==2.0
PyYAML==3.12
-e git+git@github.com:juiceinc/recipe@d71e968c66a1ba8ff6b93ba39e4f1ec09ad962d0#egg=recipe
recipe-caching==0.1.0
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.0b2
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
